+++ src/common/index.ts
@@ -15,4 +15,4 @@ export { default as Profile } from "./Profile/Profile";
 export { default as ProfileItem } from "./Profile/ProfileItem/ProfileItem";
 export { default as ProfileEdit } from "./Profile/ProfileEdit/ProfileEdit";
 export { default as CommentItem } from "./Comment/CommentItem";
-export { default as Snackbar } from "./Snackbar/UnsavedSnackbar";
+export { default as SaveChangesSnackbar } from "./SaveChangesSnackbar/SaveChangesSnackbar";
